% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create a DuckLake table}
\usage{
create_table(table_name, data_source)
}
\arguments{
\item{table_name}{Name of the new table}

\item{data_source}{Raw data source}
}
\description{
Create a DuckLake table
}
