% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{extract_assignments_from_sql}
\alias{extract_assignments_from_sql}
\title{Extract column assignments from SQL SELECT statement}
\usage{
extract_assignments_from_sql(sql_text)
}
\arguments{
\item{sql_text}{A SQL SELECT statement}
}
\value{
A string of comma-separated column assignments for UPDATE SET clause
}
\description{
Extract column assignments from SQL SELECT statement
}
\keyword{internal}
