% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ducklake_table.R
\name{get_metadata_table}
\alias{get_metadata_table}
\title{Get a DuckLake metadata table}
\usage{
get_metadata_table(tbl_name, ducklake_name)
}
\arguments{
\item{tbl_name}{Character string, name of the table to retrieve}

\item{ducklake_name}{Character string, name of the ducklake database}
}
\value{
A DuckLake table of class \code{tbl_duckdb_connection}
}
\description{
Get a DuckLake metadata table
}
