% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_path.R
\name{get_table_path}
\alias{get_table_path}
\title{Find the Parquet file path for a DuckLake table}
\usage{
get_table_path(table_name, con)
}
\arguments{
\item{table_name}{Name of the table to find the file for}

\item{con}{DuckDB connection to a DuckLake database}
}
\value{
Character string containing the Parquet file path
}
\description{
Find the Parquet file path for a DuckLake table
}
