% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ducklake.R
\name{install_ducklake}
\alias{install_ducklake}
\title{Install the ducklake extension to duckdb}
\usage{
install_ducklake()
}
\description{
This only needs to be run once on your system.
}
\examples{
ducklake::install_ducklake()
}
