% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ducklake_query.R
\name{show_ducklake_query}
\alias{show_ducklake_query}
\title{Show the SQL that would be executed by ducklake operations}
\usage{
show_ducklake_query(.data, table_name)
}
\arguments{
\item{.data}{A dplyr query object (tbl_lazy)}

\item{table_name}{The target table name for the operation}
}
\value{
The first argument, invisibly (following show_query convention)
}
\description{
This function shows the SQL that would be generated and executed by ducklake.
This is useful for debugging and understanding what SQL is being sent to DuckDB.
}
\examples{
\dontrun{
# Show SQL for an update operation
get_ducklake_table("my_table") \%>\%
  mutate(status = "updated") \%>\%
  show_ducklake_query("my_table")
}
}
