% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{update_table}
\alias{update_table}
\title{Convert a dplyr SELECT query to a DuckLake UPDATE statement}
\usage{
update_table(query, table_name, ducklake_name = "my_ducklake")
}
\arguments{
\item{query}{A dplyr query object that can be passed to show_query()}

\item{table_name}{The name of the table to update}

\item{ducklake_name}{The name of the DuckLake database, defaults to "my_ducklake"}
}
\value{
A string containing the DuckLake UPDATE statement
}
\description{
Convert a dplyr SELECT query to a DuckLake UPDATE statement
}
\examples{
\dontrun{
# Create a dplyr query
train_file |>
  mutate(
    name_long = case_when(
      code == "ASB" ~ "Johan Cruijff ArenA",
      .default = name_long
    )
  ) |>
  update_table("nl_train_stations", ducklake_name = "my_ducklake") |>
  duckplyr::db_exec()
}
}
